+++ boundlexx/api/v1/serializers/base.py
@@ -220,6 +220,8 @@ class URLSimpleItemSerializer(serializers.ModelSerializer):
             "game_id",
             "name",
             "string_id",
+            "has_colors",
+            "is_resource",
         ]